utils.ts to folder & isGuildMessage
